requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
joblib>=1.2.0
mysql-connector-python>=8.0.0
plotly>=5.13.0
matplotlib>=3.7.0
pillow>=9.4.0
xgboost>=1.7.0
requests>=2.28.0
python-dotenv>=0.19.0
streamlit-option-menu>=0.3.0
plotly-express>=0.4.0
geopy>=2.3.0
folium>=0.14.0
streamlit-folium>=0.15.0